Added two StandardOperators: SubSet and SubSetEq.
SubSetEq is mapped to the Viper subset operator.

A new pass is defined. This pass is meant for cases where rewrite rule on ADT operators depend on the type. The cases in Standardize (such as AppendSingle and PrependSingle) have been moved to this new pass as well to simplify their code.