Added a constructor for Maps.
Maps are encoded as structvalues and eventually an invokation to a domain function. Besides the domain for the map, there is also a domain for a tuple. This tuple is cannot be accessed through VerCors/PVL, however that might change in the future. StandardOperators for the functions on maps have been added.