Add capability to make domain functions with 0 arguments "unique".
This commit adds the capability to add "unique" domain functions to an ADT. This feature is only available programmatically, as there is no frontend syntax that enables this. A "unique" function is a function with 0 formal parameters which is guaranteed to yield a value different from all other unique functions in that ADT. This is useful for modeling constructors. See the viper documentation/tutorial for examples.