Issue #360 was not fully solved.
The previous solution solved concrete cases, however it did not solve more general cases. This was due to the undefined behavior of sum when you summed outside of the range. As a solution, a new function is added named vsize that for a sequence is defined. A new axiom is added stating that if the range you are trying to take is invalid (lowerbound smaller than zero, upperbound larger than the sequence), then the sum is 0.