Move the pvl-encode pass into ColPVLParser.
Initially the pvl-encode pass was in main, and besides checking if any operators are used, it also checked whether silver was the target language. Since silver is currently effectively the only backend, and the encoding step is sensible, in ColPVLParser it is not checked if silver is the target language and the encoding is always applied. If this turns out to be a bad decision the silver check can be reintroduced within ColPVLParser.