Enable return in the abrupt termination exception transformation.
When return x; is transformed into throw new return_exception(x);, the contract of the constructor of return_exception must ensure that the permission for the field "value" is returned and ensure that the field value is equal to x (i.e. the first arg). Otherwise it's hard to fulfill postconditions about the return value.