Finished Iteration merging.
Loop unrolling has been made more general to work with iteration merging as well. The implementation will have to be revisited at some point to make it a bit more efficient. Iteration merging has been implemented for the case where we do have to unroll. Type checking for Iteration merging seems to be exactly the same as loop unrolling.