Coerce return values to the return type of method.
Previously return values were not coerced to the return type of the
method. This caused "return null" not to be converted into "return
VCTNone", because the return type wasn't coerced to VCTOption, but was
just left as-is. Now return values are type-coerced to the method return
type, similarly as values that are being assigned to a variable.