Only force frac type of literals.
Previously the typechecker would force the secon argument of a Perm to be frac. This would cause the rewriter to rewrite an int variable as a frac. We narrow this to only forcing literals to be frac, and leave all other types as they are. This ensures expressions like "read" are not matched against rewrite matching variables of type int.