Reinstate monotonicBool & monotonicBoolBroken.
The rewrite rules are adjusted to be more limited in how read
permissions are rewritten. The result should be similar, except that
deep expressions with read will require more rewrite rules now. Luckily,
they are pretty rare.

The monotonicBool tests are removed from the problemFail suite, and
toSeq is replaced by the builtin \values.