Remove true and false from spec.
True and false are removed from spec and reintroduced to PVL, because having them in spec only causes some problems when converting valReserved keywords (valReserved used to only contain nameExpressions, but if true and false are added to that it also includes constants).

Because true and false are not in the spec anymore, the convertValReserved routine in CMLtoCOL can also have return type NameExpression.