Change syntax for axioms.
Syntax for axioms is changed from "langExpr == langExpr" to just "langExpr". This still accepts the same language, but the precedence changes in some cases. Therefore most of the time parentheses need to be included now.

This change avoids slow parsing, since the stray "==" that doesn't participate in normal expression parsing confuses the parser.