Test wiki snippets in CI (#714)
This commit adds functionality of testing all snippets present in the wiki when the wiki is updated, as well as including the snippets from the wiki in the regular test suite. The wiki script is extended to dump all extracted cases to a folder with proper testing headers. Finally, a github action is added which ensures in PRs the testing workflow is not run in duplo.