Rename internal types and make more narrow.
VeyMontThread is renamed to Endpoint as that is an unused, nicely
generic, covers the meaning and also occurs more often in session type
literatur.

The Endpoint type is made more specific, in that it can now only
reference Class, and not any given Type. The benefit of this is that it
now no longer requires typechecking. The downside is that it causes a
whole lot of cascading changes.