Collect bip_annotation modifier in all ToCol classes
The bip_annotation modifier is part of the specification grammar for all input languages, so a PVL program like this parses according to ANTLR:

```
bip_annotation
void foo() {
}
```

...but then hit a MatchError ("VerCors has crashed") in the PVLToCol transform.

This changes PVLToCol, CToCol and CPPToCol to follow the same strategy as LLVMContractToCol, which is to pass through the BIP annotation. In practice, this results in a "modifier cannot be attached to this declaration" message in the example above.